Fixed norm issues, ready to PR
