+++ Makefile
@@ -6,7 +6,7 @@
 #    By: tsuchen <tsuchen@student.42.fr>            +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/08/15 12:02:08 by tsuchen           #+#    #+#              #
-#    Updated: 2024/08/29 11:12:59 by tsuchen          ###   ########.fr        #
+#    Updated: 2024/08/29 12:13:13 by tsuchen          ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -28,7 +28,7 @@ SRCS_EV		=	key_events.c
 
 SRCS_MAP	= map.c
 
-SRCS_RC		= raycasting.c
+SRCS_RC		= raycasting.c render_utils.c
 
 SRCS_VEC	= vector.c vector_2.c vector_3.c
